mcp_vton/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
redis==5.0.8
httpx==0.27.2
pillow==10.4.0
google-genai==0.3.0
beautifulsoup4==4.12.3
lxml==5.3.0
python-multipart==0.0.9
pydantic==2.9.2
starlette==0.38.5
